# app/contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Zhixue Mate Agent API",
    "version": "api-contract-v0.3",
    "description": "知学 Mate 前后端唯一接口契约。字段变更必须先升契约版本并通知前端。 声明层已于 2026-09-22 按实跑结果补全：补齐 5 个已实现端点、所有 operation 的 405、register 的 409、exercises 的 4 个查询参数与 400、knowledge-bases 的 400、可空字段的 nullable、版本头参数，并移除死枚举值 LLM_FALLBACK。实现未改动。 KnowledgeBaseDetailResponse.documents 已于 2026-09-22 补声明（后端返回、前端在读，此前契约未声明）。 LegacyChatResponse.card.summary 已于 2026-09-22 补声明（前端在读、后端已补返回，此前两边都缺）。"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "本地开发"
    },
    {
      "url": "http://{host}:5000",
      "description": "局域网联调",
      "variables": {
        "host": {
          "default": "192.168.1.100"
        }
      }
    }
  ],
  "security": [],
  "paths": {
    "/api/agent/chat": {
      "post": {
        "summary": "Legacy frontend-compatible chat entry",
        "operationId": "legacyAgentChat",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegacyChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyChatResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "405": {
            "description": "请求方法不被该路径支持（后端统一返回 JSON 错误体，而非 HTML）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "可选鉴权。/api/v1/** 不带 token 时按演示/游客身份（demo-user）处理，以保证演示主链不受影响。"
      }
    },
...
    "schemas": {
      "LegacyChatRequest": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "user_message": {
            "type": "string"
          },
          "image": {
            "type": "string"
          }
        }
      },
      "LegacyChatResponse": {
        "type": "object",
        "required": [
          "reply",
          "intent",
          "card"
        ],
        "properties": {
          "reply": {
            "type": "string"
          },
          "intent": {
            "type": "string",
            "enum": [
              "unknown",
              "analyze_wrong",
              "query_tasks",
              "analyze_weakness",
              "match_partner",
              "update_profile",
              "get_suggestion"
            ]
          },
          "card": {
            "type": "object",
            "nullable": true,
            "properties": {
              "type": {
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "actionLabel": {
                "type": "string"
              },
              "targetPage": {
                "type": "string"
              },
              "sessionId": {
                "type": "string"
              },
              "summary": {
                "type": "string",
                "description": "卡片的一行描述，由前端渲染在卡片标题下方。后端 `build_card()` 按意图给出；缺失会导致描述行空白。"
              }
            },
            "description": "意图对应的结构化卡片；targetPage 必须是 App 已注册路由。"
          },
          "llmUsed": {
            "type": "boolean",
            "description": "true=大模型生成；false=本地确定性规则兜底（未配置 Key 或调用失败）。前端 ChatMain 据此显示来源标签。"
          },
          "wrongAnalysis": {
            "type": "object",
            "nullable": true,
            "properties": {
              "knowledgePoint": {
                "type": "string"
              },
              "wrongType": {
                "type": "string"
              }
            },
            "description": "仅 analyze_wrong 意图下可能非空：错题分析识别出的**结构化**知识点。前端据此把「拍一道错题」接到「针对这个知识点练」——知识点若只存在于自然语言答复里，程序无法使用。识别不出时不下发该字段，前端如实回落，不编造知识点。"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "errorCode",
          "message"
        ],
        "properties": {
          "errorCode": {
            "type": "string",
            "enum": [
              "BAD_REQUEST",
              "NOT_FOUND",
              "CONFLICT",
              "VALIDATION_ERROR",
              "UNAUTHORIZED",
              "CONTRACT_VERSION_MISMATCH",
              "INTERNAL_ERROR",
              "RATE_LIMITED",
              "DAILY_LIMIT_REACHED",
              "PHONE_LOCKED",
              "INVALID_CODE",
              "CODE_EXPIRED",
              "CODE_NOT_FOUND",
              "SMS_NOT_CONFIGURED",
              "SMS_PROVIDER_UNAVAILABLE"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "nullable": true
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  }
}